Move table component
